+++ src/main/java/dao/GameDao.java
@@ -2,7 +2,7 @@ package dao;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import json.ResultDto;
+import domain.dto.ResultDto;
 
 import java.sql.*;
 import java.util.ArrayList;
@@ -11,11 +11,11 @@ import java.util.List;
 public class GameDao {
     public Connection getConnection() {
         Connection con = null;
-        String server = "localhost:13306"; // MySQL 서버 주소
-        String database = "db_name"; // MySQL DATABASE 이름
+        String server = "localhost:13306";
+        String database = "db_name";
         String option = "?useSSL=false&serverTimezone=UTC";
-        String userName = "root"; //  MySQL 서버 아이디
-        String password = "root"; // MySQL 서버 비밀번호
+        String userName = "root";
+        String password = "root";
 
         try {
             Class.forName("com.mysql.cj.jdbc.Driver");
