+++ src/main/java/domain/dto/ResultDto.java
@@ -1,4 +1,4 @@
-package json;
+package domain.dto;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
